invoice-ocr-system/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
openai==1.3.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
databases==0.8.0
aiosqlite==0.19.0
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.1
email-validator==2.1.0